fixed logic

+++ venue-service/src/main/java/com/venuex/event_service/controller/VenueController.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Map;
 
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 import org.springframework.http.HttpStatus;
 import com.venuex.event_service.entities.SeatSection;
 import com.venuex.event_service.entities.Venue;
-import com.venuex.event_service.service.VenueService;;
+import com.venuex.event_service.service.VenueService;
+
+import jakarta.servlet.http.HttpServletRequest;;
 
 @RestController
 @RequestMapping("/api")
@@ -31,37 +34,61 @@ public class VenueController {
     //ADMINS only 
     @PostMapping("/admin/venues")
     @ResponseStatus(HttpStatus.CREATED)
-    public Venue addVenue(@RequestBody Venue venue) {
+    public Venue addVenue(@RequestBody Venue venue, HttpServletRequest request) {
+        String requesterRole = request.getHeader("X-User-Role");
+        if (!requesterRole.equals("ADMIN")) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         return venueService.createVenue(venue);
     }
 
     @PutMapping("/admin/venues/{id}")
     @ResponseStatus(HttpStatus.OK)
-    public Venue updateVenue(@PathVariable Integer id, @RequestBody Venue venue) {
+    public Venue updateVenue(@PathVariable Integer id, @RequestBody Venue venue, HttpServletRequest request) {
+        String requesterRole = request.getHeader("X-User-Role");
+        if (!requesterRole.equals("ADMIN")) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         return venueService.updateVenue(id, venue);
     }
 
     @DeleteMapping("/admin/venues/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteVenue(@PathVariable Integer id) {
+    public void deleteVenue(@PathVariable Integer id, HttpServletRequest request) {
+        String requesterRole = request.getHeader("X-User-Role");
+        if (!requesterRole.equals("ADMIN")) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         venueService.deleteVenue(id);
     }
 
     //Seat Section Operations 
      @GetMapping("/venues/{venueId}/seat-sections")
-    public List<SeatSection> getVenueSeatSections(@PathVariable Integer venueId) {
+    public List<SeatSection> getVenueSeatSections(@PathVariable Integer venueId, HttpServletRequest request) {
+        String requesterRole = request.getHeader("X-User-Role");
+        if (!requesterRole.equals("ADMIN")) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         return venueService.getVenueSeatSections(venueId);
     }
 
     @PostMapping("/admin/venues/{venueId}/seat-sections")
     @ResponseStatus(HttpStatus.CREATED)
-    public List<SeatSection> createSeatSections(@PathVariable Integer venueId, @RequestBody List<SeatSection> sections) {
+    public List<SeatSection> createSeatSections(@PathVariable Integer venueId, @RequestBody List<SeatSection> sections, HttpServletRequest request) {
+        String requesterRole = request.getHeader("X-User-Role");
+        if (!requesterRole.equals("ADMIN")) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         return venueService.createSeatSections(venueId, sections);
     }
 
     @PutMapping("/admin/venues/{venueId}/seat-sections")
     @ResponseStatus(HttpStatus.OK)
-    public List<SeatSection> updateSeatSections(@PathVariable Integer venueId, @RequestBody Map<String, Integer> sections) {
+    public List<SeatSection> updateSeatSections(@PathVariable Integer venueId, @RequestBody Map<String, Integer> sections, HttpServletRequest request) {
+        String requesterRole = request.getHeader("X-User-Role");
+        if (!requesterRole.equals("ADMIN")) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED);
+        }
         return venueService.updateSeatSections(venueId, sections);
     }
 }
